Move ReferenceListWidget into its own file for readability and reusability
